CI/CD checks
Add measures to check basic health of the application with each push to production.
Check that both production and development modes build with each push to ensure that bugs and linting errors do not pile up with new features/refactorings. Check that the API and frontends are reachable on ports 8000 and 3000, respectively